refactor(logic): relocate naval_resolution world->turn (Refs #3290)
Move naval_resolution.dart (+ its part fragments) out of the leaf
world/ layer into turn/, eliminating the world->combat and world->dossier
wrong-direction edges enumerated in #3290 Phase-0. naval_resolution
orchestrates naval combat (combat/naval_combat_resolver.dart) and
dossier/dialogue side-effects, so it belongs above the colonizethis_world
leaf layer; turn already depends on combat, diplomacy/dossier, and world.

Leaf-layer world/fog_resolution.dart now imports the coastal-visibility
helpers directly from world/naval_coastal_visibility.dart instead of via
the relocated library's re-export. Behaviour-preserving structural move.
