+++ packages/colonizethis_logic/lib/src/turn/naval_resolution_helpers.dart
