+++ packages/colonizethis_logic/test/naval_behavior_scenarios_test.dart
@@ -1,5 +1,5 @@
 import 'package:colonizethis_logic/colonizethis_logic.dart' show NavalCombatResultEvent;
-import 'package:colonizethis_logic/src/world/naval_resolution.dart';
+import 'package:colonizethis_logic/src/turn/naval_resolution.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
