+++ packages/colonizethis_logic/test/naval_test_part1_ship_reveal_end_of_turn_test.dart
@@ -1,7 +1,7 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_logic/src/turn/end_of_turn_resolver.dart';
-import 'package:colonizethis_logic/src/world/naval_resolution.dart';
+import 'package:colonizethis_logic/src/turn/naval_resolution.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
